+++ lib/constants/constants.dart
@@ -1,0 +1,5 @@
+class Constants {
+  static const int colorBackgroundTop = 0xFFF44336;
+  static const int colorBackgroundBottom = 0xFF212133;
+  static const int colorLabel = 0xCDFFFBFB;
+}
